Tidy up tests for 'source' rule
